Move DocIndexIterator logic to GraphNodeIdToDocMap
